+++ src/universalelectricity/api/energy/IEnergyContainer.java
@@ -14,15 +14,15 @@ public interface IEnergyContainer
 	 * 
 	 * This function is NOT recommended for calling.
 	 */
-	public void setEnergy(int energy);
+	public void setEnergy(long energy);
 
 	/**
 	 * * @return Get the amount of energy currently stored in the block.
 	 */
-	public int getEnergy(ForgeDirection from);
+	public long getEnergy(ForgeDirection from);
 
 	/**
 	 * @return Get the max amount of energy that can be stored in the block.
 	 */
-	public int getEnergyCapacity(ForgeDirection from);
+	public long getEnergyCapacity(ForgeDirection from);
 }
